Change id to long.

+++ Models/Bet.cs
@@ -2,15 +2,15 @@ namespace BetMe.Models;
 
 public class Bet
 {
-    public int Id { get; set; } = 0;
+    public long Id { get; set; } = 0;
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public int CreatorId { get; set; }
+    public long CreatorId { get; set; }
     public DateTime CreatedAt { get; set; }
     public BetStatus Status { get; set; }
-    public int WinOutcomeId { get; set; } = -1;
+    public long WinOutcomeId { get; set; } = -1;
 
-    public Bet(string name, string description, int creatorId)
+    public Bet(string name, string description, long creatorId)
     {
         Name = name;
         Description = description;
