+++ Models/DTO/OutcomeDto.cs
@@ -7,5 +7,5 @@ public class OutcomeDto
     [Required, MaxLength(256)]
     public String Name { get; set; } = string.Empty;
     [Required]
-    public int BetId { get; set; }
+    public long BetId { get; set; }
 }
