+++ Models/Outcome.cs
@@ -2,9 +2,9 @@ namespace BetMe.Models;
 
 public class Outcome
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public String Name { get; set; } = string.Empty;
-    public int BetId { get; set; }
+    public long BetId { get; set; }
     public int Selections { get; set; } = 0;
     public int Votes { get; set; } = 0;
 }
