+++ Models/User.cs
@@ -2,7 +2,7 @@ namespace BetMe.Models;
 
 public class User
 {
-    public int Id { get; set; } = 0;
+    public long Id { get; set; } = 0;
     public string Name { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public string PasswordHash { get; set; } = string.Empty;
