+++ Models/UserBet.cs
@@ -2,10 +2,10 @@ namespace BetMe.Models
 {
     public class UserBet
     {
-        public int Id { get; set; }
-        public int BetId { get; set; }
-        public int UserId { get; set; }
-        public int OutcomeId { get; set; }
+        public long Id { get; set; }
+        public long BetId { get; set; }
+        public long UserId { get; set; }
+        public long OutcomeId { get; set; }
         public bool HasVoted { get; set; } = false;
     }
 }
