+++ Services/IBetService.cs
@@ -14,20 +14,20 @@ public interface IBetService
     /// </summary>
     /// <param name="id">Id of finding bet.</param>
     /// <returns> Bet with provided id.</returns>
-    Task<Bet> GetBetByIdAsync(int id);
+    Task<Bet> GetBetByIdAsync(long id);
     /// <summary>
     /// Gets all participants of the bet.
     /// </summary>
     /// <param name="betId"> Id of the bet.</param>
     /// <returns>List of participants.</returns>
-    Task<List<UserBet>> GetAllUsersOfBetAsync(int betId);
+    Task<List<UserBet>> GetAllUsersOfBetAsync(long betId);
     /// <summary>
     /// Creates a new bet.
     /// </summary>
     /// <param name="bet"> Bet params.</param>
     /// <param name="userId"> Creator. </param>
     /// <returns> Created bet.</returns>
-    Task<Bet> CreateBetAsync(BetCreatingDto bet, int userId);
+    Task<Bet> CreateBetAsync(BetCreatingDto bet, long userId);
     /// <summary>
     /// Adds a participant to the bet.
     /// </summary>
@@ -39,11 +39,11 @@ public interface IBetService
     /// </summary>
     /// <param name="betId"> Bet to start.</param>
     /// <returns> Started bet.</returns>
-    Task<Bet> StartBetAsync(int betId);
+    Task<Bet> StartBetAsync(long betId);
     /// <summary>
     /// Deletes the bet.
     /// </summary>
     /// <param name="betId"> Id of bet to delete.</param>
     /// <returns> Deleted bet.</returns>
-    Task<Bet> DeleteBetAsync(int betId);
+    Task<Bet> DeleteBetAsync(long betId);
 }
